requirements.txt
librosa
numpy
soundfile
pydub
torch
pyannote.audio
speechbrain
scikit-learn
pandas
tqdm
PyYAML
Flask
werkzeug
joblib
psutil
